A little bit of refactoring and m/b even future compatibility with frames
(if eloraam releases any). Going to remove wireless conductors, or maybe
nerf them a bit. Synchronizer will have a quite difficult recipe
(difficult, but not expensive). Should be compatible with anything that
moves things (in current World, as it checks for it). It's quite
possible that you will understand none of this rambling..